Run tests only with session fixtures [WIP]
