+++ tests/test_analysis.py
