+++ tests/test_deletions.py
