+++ tests/test_forms_climbing_sessions.py
@@ -7,8 +7,7 @@ from datetime import datetime
 
 from sqlalchemy import text
 
-from .conftest import CLIMBER_ID, SLEEP_TIME
-from .form_utils import start_session, stop_session, started_session_id
+from .conftest import CLIMBER_ID, SLEEP_TIME, start_session, stop_session
 
 
 EXISTING_OBJECTS = {
@@ -50,35 +49,10 @@ def test_create_invalid_session(driver) -> None:
         assert not form_accepted, data
 
 
-def test_create_session_on_existing_area(db_session, started_session_id) -> None:
-    """
-    Start a session on an existing area.
-    """
-
-    # check that the session was created
-    sql_query = text(
-        f"""
-        SELECT id FROM climbing_session
-        WHERE date = '{NEW_OBJECTS["session_date"].strftime("%Y-%m-%d")}'
-        AND climber_id = {CLIMBER_ID}
-        AND area_id = (SELECT id FROM area WHERE name = '{EXISTING_OBJECTS["area"]}');
-        """
-    )
-    results = db_session.execute(sql_query).fetchall()
-    assert len(results) == 1
-
-    # check that the number of areas in the database did not change
-    n_areas_after = db_session.execute(text("SELECT COUNT(*) FROM area")).fetchall()[0][
-        0
-    ]
-    assert n_areas_after == 2
-
-
 def test_create_session_on_new_area(driver, db_session) -> None:
     """
     Start a session on a new area.
     """
-    stop_session(driver)
 
     # get the current number of areas in the database
     n_areas_query = text("SELECT COUNT(*) FROM area")
@@ -112,3 +86,28 @@ def test_create_session_on_new_area(driver, db_session) -> None:
     n_areas_after = db_session.execute(n_areas_query).fetchall()[0][0]
     assert n_areas_after == n_areas_before + 1
     stop_session(driver)
+
+
+def test_create_session_on_existing_area(db_session, started_session_id) -> None:
+    """
+    Start a session on an existing area.
+    """
+
+    # check that the session was created
+    sql_query = text(
+        f"""
+        SELECT id FROM climbing_session
+        WHERE date = '{NEW_OBJECTS["session_date"].strftime("%Y-%m-%d")}'
+        AND climber_id = {CLIMBER_ID}
+        AND area_id = (SELECT id FROM area WHERE name = '{EXISTING_OBJECTS["area"]}');
+        """
+    )
+    results = db_session.execute(sql_query).fetchall()
+    assert len(results) == 1
+
+    # check that the number of areas in the database did not change
+    # i.e. 2 if this test is run alone, 3 if the previous test is run before
+    n_areas_after = db_session.execute(text("SELECT COUNT(*) FROM area")).fetchall()[0][
+        0
+    ]
+    assert n_areas_after in [2, 3]
