+++ tests/test_forms_videos.py
@@ -15,8 +15,13 @@ from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from sqlalchemy import text
 
-from .conftest import HOME_TITLE, HOME_URL, EXISTING_OBJECTS
-from .form_utils import get_existing_route, fill_form, started_session_id
+from .conftest import (
+    HOME_TITLE,
+    HOME_URL,
+    EXISTING_OBJECTS,
+    get_existing_route,
+    fill_form,
+)
 import subprocess
